+++ src/deck-components/Router.tsx
@@ -1,3 +1,4 @@
+import { sleep } from '../utils';
 import { Module, findModuleChild } from '../webpack';
 
 export enum SideMenu {
@@ -125,32 +126,65 @@ export interface Navigation {
   CloseSideMenus(): void;
 }
 
-export const Navigation = {
-  Navigate: Router.Navigate.bind(Router),
-  NavigateBack: Router.WindowStore?.GamepadUIMainWindowInstance?.NavigateBack.bind(
-    Router.WindowStore.GamepadUIMainWindowInstance,
-  ),
-  NavigateToAppProperties: Router.NavigateToAppProperties.bind(Router),
-  NavigateToExternalWeb: Router.NavigateToExternalWeb.bind(Router),
-  NavigateToInvites: Router.NavigateToInvites.bind(Router),
-  NavigateToChat: Router.NavigateToChat.bind(Router),
-  NavigateToLibraryTab: Router.NavigateToLibraryTab.bind(Router),
-  NavigateToLayoutPreview: Router.NavigateToLayoutPreview.bind(Router),
-  NavigateToSteamWeb: Router.WindowStore?.GamepadUIMainWindowInstance?.NavigateToSteamWeb.bind(
-    Router.WindowStore.GamepadUIMainWindowInstance,
-  ),
-  NavigateToWebRoute: Router.WindowStore?.GamepadUIMainWindowInstance?.NavigateToWebRoute.bind(
-    Router.WindowStore.GamepadUIMainWindowInstance,
-  ),
-  OpenSideMenu: Router.WindowStore?.GamepadUIMainWindowInstance?.MenuStore.OpenSideMenu.bind(
-    Router.WindowStore.GamepadUIMainWindowInstance.MenuStore,
-  ),
-  OpenQuickAccessMenu: Router.WindowStore?.GamepadUIMainWindowInstance?.MenuStore.OpenQuickAccessMenu.bind(
-    Router.WindowStore.GamepadUIMainWindowInstance.MenuStore,
-  ),
-  OpenMainMenu: Router.WindowStore?.GamepadUIMainWindowInstance?.MenuStore.OpenMainMenu.bind(
-    Router.WindowStore.GamepadUIMainWindowInstance.MenuStore,
-  ),
-  CloseSideMenus: Router.CloseSideMenus.bind(Router),
-  OpenPowerMenu: Router.OpenPowerMenu.bind(Router),
-} as Navigation;
+export let Navigation = {} as Navigation;
+
+try {
+  (async () => {
+    let InternalNavigators: any = {};
+    if (!Router.NavigateToAppProperties || (Router as unknown as any).deckyShim) {
+      function initInternalNavigators() {
+        try {
+          InternalNavigators = findModuleChild((m: any) => {
+            if (typeof m !== 'object') return undefined;
+            for (let prop in m) {
+              if (m[prop]?.GetNavigator) {
+                return m[prop];
+              }
+            }
+          })?.GetNavigator();
+        } catch (e) {
+          console.error('[DFL:Router]: Failed to init internal navigators, trying again');
+        }
+      }
+      initInternalNavigators();
+      while (!InternalNavigators?.AppProperties) {
+        console.log('[DFL:Router]: Trying to init internal navigators again');
+        await sleep(100);
+        initInternalNavigators();
+      }
+    }
+    const newNavigation = {
+      Navigate: Router.Navigate.bind(Router),
+      NavigateBack: Router.WindowStore?.GamepadUIMainWindowInstance?.NavigateBack.bind(
+        Router.WindowStore.GamepadUIMainWindowInstance,
+      ),
+      NavigateToAppProperties: InternalNavigators?.AppProperties || Router.NavigateToAppProperties.bind(Router),
+      NavigateToExternalWeb: InternalNavigators?.ExternalWeb || Router.NavigateToExternalWeb.bind(Router),
+      NavigateToInvites: InternalNavigators?.Invites || Router.NavigateToInvites.bind(Router),
+      NavigateToChat: Router.NavigateToChat.bind(Router),
+      NavigateToLibraryTab: InternalNavigators?.LibraryTab || Router.NavigateToLibraryTab.bind(Router),
+      NavigateToLayoutPreview: Router.NavigateToLayoutPreview.bind(Router),
+      NavigateToSteamWeb: Router.WindowStore?.GamepadUIMainWindowInstance?.NavigateToSteamWeb.bind(
+        Router.WindowStore.GamepadUIMainWindowInstance,
+      ),
+      NavigateToWebRoute: Router.WindowStore?.GamepadUIMainWindowInstance?.NavigateToWebRoute.bind(
+        Router.WindowStore.GamepadUIMainWindowInstance,
+      ),
+      OpenSideMenu: Router.WindowStore?.GamepadUIMainWindowInstance?.MenuStore.OpenSideMenu.bind(
+        Router.WindowStore.GamepadUIMainWindowInstance.MenuStore,
+      ),
+      OpenQuickAccessMenu: Router.WindowStore?.GamepadUIMainWindowInstance?.MenuStore.OpenQuickAccessMenu.bind(
+        Router.WindowStore.GamepadUIMainWindowInstance.MenuStore,
+      ),
+      OpenMainMenu: Router.WindowStore?.GamepadUIMainWindowInstance?.MenuStore.OpenMainMenu.bind(
+        Router.WindowStore.GamepadUIMainWindowInstance.MenuStore,
+      ),
+      CloseSideMenus: Router.CloseSideMenus.bind(Router),
+      OpenPowerMenu: Router.OpenPowerMenu.bind(Router),
+    } as Navigation;
+
+    Object.assign(Navigation, newNavigation);
+  })();
+} catch (e) {
+  console.error('[DFL:Router]: Error initializing Navigation interface', e);
+}